+++ app/controllers/forest_liana/resources_controller.rb
@@ -15,6 +15,15 @@ module ForestLiana
 
     def index
       begin
+
+        if request.format == 'csv'
+          return head :forbidden if !PermissionsChecker.new(@resource, "export").is_authorized
+        elsif params.has_key?(:searchToEdit)
+          return head :forbidden if !PermissionsChecker.new(@resource, "searchToEdit").is_authorized
+        else
+          return head :forbidden if !PermissionsChecker.new(@resource, "list").is_authorized
+        end
+
         getter = ForestLiana::ResourcesGetter.new(@resource, params)
         getter.perform
 
@@ -30,6 +39,8 @@ module ForestLiana
 
     def show
       begin
+        return head :forbidden if !PermissionsChecker.new(@resource, "show").is_authorized
+
         getter = ForestLiana::ResourceGetter.new(@resource, params)
         getter.perform
 
@@ -43,6 +54,8 @@ module ForestLiana
 
     def create
       begin
+        return head :forbidden if !PermissionsChecker.new(@resource, "create").is_authorized
+
         creator = ForestLiana::ResourceCreator.new(@resource, params)
         creator.perform
 
@@ -64,6 +77,8 @@ module ForestLiana
 
     def update
       begin
+        return head :forbidden if !PermissionsChecker.new(@resource, "update").is_authorized
+
         updater = ForestLiana::ResourceUpdater.new(@resource, params)
         updater.perform
 
@@ -85,6 +100,8 @@ module ForestLiana
 
     def destroy
       begin
+        return head :forbidden if !PermissionsChecker.new(@resource, "delete").is_authorized
+
         @resource.destroy(params[:id])
         head :no_content
       rescue => error