+++ app/services/forest_liana/query_stat_getter.rb
@@ -1,7 +1,5 @@
 module ForestLiana
   class QueryStatGetter
-    QUERY_SELECT = /\ASELECT\s.*FROM\s.*\z/im
-
     attr_accessor :record
 
     def initialize(params)
@@ -11,7 +9,7 @@ module ForestLiana
     def perform
       raw_query = @params['query'].strip
 
-      check_query(raw_query)
+      LiveQueryChecker.new(raw_query, 'Live Query Chart').validate()
 
       if @params['record_id']
         raw_query.gsub!('?', @params['record_id'].to_s)
@@ -68,14 +66,6 @@ module ForestLiana
     end
 
     private
-
-    def check_query(query)
-      raise 'You cannot execute an empty SQL query.' if query.blank?
-      if query.include?(';') && query.index(';') < (query.length - 1)
-        raise 'You cannot chain SQL queries.'
-      end
-      raise 'Only SELECT queries are allowed.' if QUERY_SELECT.match(query).nil?
-    end
 
     def error_message(result, key_names)
       "The result columns must be named #{key_names} instead of '#{result.keys.join("', '")}'"