+++ app/services/forest_liana/resources_getter.rb
@@ -19,6 +19,22 @@ module ForestLiana
         @records = @records.where(@segment.where.call())
       end
 
+      # NOTICE: Live Query mode
+      if @params[:segmentQuery]
+        LiveQueryChecker.new(@params[:segmentQuery], 'Live Query Segment').validate()
+
+        begin
+          results = ActiveRecord::Base.connection.execute(@params[:segmentQuery])
+        rescue => error
+          error_message = "Live Query Segment: #{error.message}"
+          FOREST_LOGGER.error(error_message)
+          raise ForestLiana::Errors::LiveQueryError.new(error_message)
+        end
+
+        record_ids = results.to_a.map { |record| record['id'] }
+        @records = @records.where(id: record_ids)
+      end
+
       @records = search_query
       @records_to_count = @records