+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Change Log
 
 ## [Unreleased]
+### Added
+- Live Query - Let users define segments using SQL.
+-
 
 ## RELEASE 2.6.1 - 2018-05-29
 ### Fixed